adding user input helper (with autocomplete)
